Fixed drag and drop image issue when outside editor
